amplify-xcode: refactor CLICommands into separate files
